+++ protocol/event_test.go
@@ -5,6 +5,8 @@ import "testing"
 func TestEventMarkers(t *testing.T) {
 	(&StreamOpenedEvent{}).isEvent()
 	(&StreamFeaturesEvent{}).isEvent()
+	(&StartTLSProceedEvent{}).isEvent()
+	(&StartTLSFailureEvent{}).isEvent()
 	(&StanzaEvent{}).isEvent()
 	(&ElementEvent{}).isEvent()
 	(&StreamClosedEvent{}).isEvent()